requirements.txt
beautifulsoup4==4.11.2
discordwebhook==1.0.3
Requests==2.31.0
python-dotenv